+++ openyich-framework/src/main/java/com/openyich/framework/boot/config/SpringRedisConfiguration.java
@@ -1,0 +1,60 @@
+package com.openyich.framework.boot.config;
+
+import java.net.UnknownHostException;
+
+import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
+import org.springframework.boot.autoconfigure.data.redis.RedisProperties;
+import org.springframework.boot.context.properties.EnableConfigurationProperties;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.redis.connection.RedisConnectionFactory;
+import org.springframework.data.redis.core.RedisOperations;
+import org.springframework.data.redis.core.RedisTemplate;
+import org.springframework.data.redis.core.StringRedisTemplate;
+import org.springframework.data.redis.serializer.Jackson2JsonRedisSerializer;
+import org.springframework.data.redis.serializer.StringRedisSerializer;
+
+import com.fasterxml.jackson.annotation.JsonAutoDetect;
+import com.fasterxml.jackson.annotation.PropertyAccessor;
+import com.fasterxml.jackson.databind.ObjectMapper;
+
+@Configuration
+@ConditionalOnClass(RedisOperations.class)
+@EnableConfigurationProperties(RedisProperties.class)
+public class SpringRedisConfiguration {
+
+  @Bean
+  @ConditionalOnMissingBean(name = "redisTemplate")
+  public RedisTemplate<Object, Object> redisTemplate(RedisConnectionFactory redisConnectionFactory)
+      throws UnknownHostException {
+    RedisTemplate<Object, Object> template = new RedisTemplate<>();
+    template.setConnectionFactory(redisConnectionFactory);
+
+    StringRedisSerializer stringRedisSerializer = new StringRedisSerializer();
+    Jackson2JsonRedisSerializer<Object> jackson2JsonRedisSerializer =
+        new Jackson2JsonRedisSerializer<>(Object.class);
+    ObjectMapper objectMapper = new ObjectMapper();
+    objectMapper.setVisibility(PropertyAccessor.ALL, JsonAutoDetect.Visibility.DEFAULT);
+    objectMapper.enableDefaultTyping(ObjectMapper.DefaultTyping.NON_FINAL);
+    jackson2JsonRedisSerializer.setObjectMapper(objectMapper);
+
+    template.setKeySerializer(stringRedisSerializer);
+    template.setValueSerializer(jackson2JsonRedisSerializer);
+
+    template.setHashKeySerializer(stringRedisSerializer);
+    template.setHashValueSerializer(jackson2JsonRedisSerializer);
+    template.afterPropertiesSet();
+    return template;
+  }
+
+  @Bean
+  @ConditionalOnMissingBean(StringRedisTemplate.class)
+  public StringRedisTemplate stringRedisTemplate(RedisConnectionFactory redisConnectionFactory)
+      throws UnknownHostException {
+    StringRedisTemplate template = new StringRedisTemplate();
+    template.setConnectionFactory(redisConnectionFactory);
+    return template;
+  }
+
+}